requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.3
httpx==0.25.2
authlib==1.2.1 